APIO-324 Split tests into client-test, client and test-base
