+++ apio-architect-client-test/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Apio Architect Client Test
+Bundle-SymbolicName: com.liferay.apio.architect.client.test
+Bundle-Version: 1.0.0
+Test-Cases: ${classes;CONCRETE;ANNOTATED;org.junit.Test}
