+++ apio-architect-sample-client-test/logback.xml
@@ -1,0 +1,14 @@
+<configuration>
+	<contextListener class="ch.qos.logback.classic.jul.LevelChangePropagator">
+		<resetJUL>true</resetJUL>
+	</contextListener>
+	<appender class="ch.qos.logback.core.ConsoleAppender" name="STDERR">
+		<target>System.err</target>
+		<encoder>
+			<pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+		</encoder>
+	</appender>
+	<root level="OFF">
+		<appender-ref ref="STDERR" />
+	</root>
+</configuration>
