+++ apio-architect-test-base/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Apio Architect Test Base
+Bundle-SymbolicName: com.liferay.apio.architect.test.base
+Bundle-Version: 1.0.0
