+++ apio-architect-test-base/build.gradle
@@ -1,0 +1,16 @@
+dependencies {
+	compileOnly group: "io.vavr", name: "vavr", version: "0.9.3"
+	compileOnly group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
+	compileOnly group: "org.apache.aries.spec", name: "org.apache.aries.javax.jax.rs-api", version: "1.0.1"
+	compileOnly group: "org.apache.felix", name: "org.apache.felix.scr", version: "2.0.8"
+	compileOnly group: "org.json", name: "json", version: "20180813"
+	compileOnly group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.jaxrs", version: "1.0.0"
+	compileOnly group: "org.osgi", name: "osgi.enroute.hamcrest.wrapper", version: "1.3.0"
+	compileOnly group: "org.osgi", name: "osgi.enroute.junit.wrapper", version: "4.12.0"
+}
+
+deploy {
+	enabled = false
+}
