--- apio-architect-test/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Apio Architect Test
-Bundle-SymbolicName: com.liferay.apio.architect.test
-Bundle-Version: 1.0.0
-Test-Cases: ${classes;CONCRETE;ANNOTATED;org.junit.Test}
