--- apio-architect-test/test.bndrun
@@ -1,61 +1,0 @@
--include: -local.bnd
-
--resolve.effective: resolve, active
-
--runblacklist:\
-	osgi.identity;filter:='(osgi.identity=org.osgi.compendium)',\
-    	osgi.identity;filter:='(osgi.identity=org.osgi.service.cm)',\
-        osgi.identity;filter:='(osgi.identity=org.osgi.service.http.whiteboard)',\
-        osgi.identity;filter:='(osgi.identity=org.apache.felix.http.api)',\
-        osgi.identity;filter:='(osgi.identity=osgi.cmpn)'
-
--runbundles: \
-	ch.qos.logback.classic;version='[1.2.3,1.2.4)',\
-	ch.qos.logback.core;version='[1.2.3,1.2.4)',\
-	com.liferay.apio.architect.api;version='[2.0.0,2.0.1)',\
-	com.liferay.apio.architect.exception.mapper.impl;version='[2.0.0,2.0.1)',\
-	com.liferay.apio.architect.impl;version='[2.0.0,2.0.1)',\
-	com.liferay.apio.architect.sample;version='[2.0.0,2.0.1)',\
-	com.liferay.apio.architect.test;version='[1.0.0,1.0.1)',\
-	com.liferay.apio.architect.test.fragment;version='[1.0.0,1.0.1)',\
-	com.liferay.apio.architect.uri.mapper.impl;version='[2.0.0,2.0.1)',\
-	io.vavr;version='[0.9.3,0.9.4)',\
-	javax.validation.api;version='[2.0.0,2.0.1)',\
-	json;version='[20180813.0.0,20180813.0.1)',\
-	org.apache.aries.javax.jax.rs-api;version='[1.0.1,1.0.2)',\
-	org.apache.aries.jax.rs.whiteboard;version='[1.0.1,1.0.2)',\
-	org.apache.felix.configadmin;version='[1.8.14,1.8.15)',\
-	org.apache.felix.eventadmin;version='[1.4.8,1.4.9)',\
-	org.apache.felix.http.jetty;version='[3.4.0,3.4.1)',\
-	org.apache.felix.http.servlet-api;version='[1.1.2,1.1.3)',\
-	org.apache.felix.scr;version='[2.0.8,2.0.9)',\
-	org.apache.servicemix.specs.annotation-api-1.3;version='[1.3.0,1.3.1)',\
-	org.osgi.service.jaxrs;version='[1.0.0,1.0.1)',\
-	org.osgi.util.function;version='[1.1.0,1.1.1)',\
-	org.osgi.util.promise;version='[1.1.0,1.1.1)',\
-	osgi.enroute.hamcrest.wrapper;version='[1.3.0,1.3.1)',\
-	osgi.enroute.junit.wrapper;version='[4.12.0,4.12.1)',\
-	slf4j.api;version='[1.7.25,1.7.26)'
-
--runee: JavaSE-1.8
-
--runfw: org.eclipse.osgi;version=3.13.0
-
--runproperties: \
-	logback.configurationFile=file:${.}/logback.xml,\
-	osgi.console.enable.builtin=false
-
--runrequires: \
-    osgi.identity;filter:='(osgi.identity=org.apache.aries.jax.rs.whiteboard)',\
-    osgi.identity;filter:='(osgi.identity=com.liferay.apio.architect.exception.mapper.impl)',\
-    osgi.identity;filter:='(osgi.identity=com.liferay.apio.architect.impl)',\
-    osgi.identity;filter:='(osgi.identity=com.liferay.apio.architect.sample)',\
-    osgi.identity;filter:='(osgi.identity=com.liferay.apio.architect.test)',\
-    osgi.identity;filter:='(osgi.identity=com.liferay.apio.architect.test.fragment)',\
-    osgi.identity;filter:='(osgi.identity=com.liferay.apio.architect.uri.mapper.impl)',\
-    osgi.identity;filter:='(osgi.identity=ch.qos.logback.classic)',\
-    osgi.identity;filter:='(osgi.identity=json)'
-
--runsystemcapabilities: ${native_capability}
-
--runvm: -javaagent:${.}/build/jacoco/jacocoagent.jar=destfile=${.}/build/jacoco/testIntegration.exec
